refactored /providers into /contexts
