+++ src/app/layout.jsx
@@ -1,7 +1,7 @@
 import { Inter } from 'next/font/google';
 import './globals.css';
-import AppSessionProvider from '@/providers/AppSessionProvider';
-import { AppDataProvider } from '@/providers/AppDataProvider';
+import { AppSessionContextProvider } from '@/contexts/AppSessionContext';
+import { AppDataContextProvider } from '@/contexts/AppDataContext';
 import Script from 'next/script';
 import { Suspense } from 'react';
 import LoadingPage from '@/components/LoadingPage';
@@ -32,13 +32,13 @@ export default function RootLayout({ children }) {
                 {/* force version update */}
                 <VersionGuard />
         
-                <AppSessionProvider>
-                    <AppDataProvider>
+                <AppSessionContextProvider>
+                    <AppDataContextProvider>
                         <Suspense fallback={<LoadingPage />}>
                             {children}
                         </Suspense>
-                    </AppDataProvider>
-                </AppSessionProvider>
+                    </AppDataContextProvider>
+                </AppSessionContextProvider>
                 <Script
                     src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/js/bootstrap.bundle.min.js"
                     strategy="afterInteractive"
