+++ src/hooks/useAlert.js
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import AlertContext from '@/contexts/AlertContext';
+import { AlertContext } from '@/contexts/AlertContext';
 
 /**
  * wrapper hook for useContext(AlertContext)
